Run distributed Hive join tests with and without dynamic filtering
Cherry pick from https://github.com/prestosql/presto/pull/1686/commits/c070ca8b4943cb5b0542651ff120c22bda192354
